code cleanup and bugfix for abort downloading not working
--HG--
extra : convert_revision : svn%3A48930e82-f9f0-11de-8173-5df4dc7aba59/trunk%4026

+++ Seven Update/Pages/Options.xaml.cs
@@ -2,7 +2,7 @@
 
 // Copyright 2007, 2008 Robert Baker, aka Seven ALive.
 // This file is part of Seven Update.
-// 
+//  
 //     Seven Update is free software: you can redistribute it and/or modify
 //     it under the terms of the GNU General Public License as published by
 //     the Free Software Foundation, either version 3 of the License, or
@@ -12,9 +12,9 @@
 //     but WITHOUT ANY WARRANTY; without even the implied warranty of
 //     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 //     GNU General Public License for more details.
-// 
+//  
 //    You should have received a copy of the GNU General Public License
-//     along with Seven Update.  If not, see <http://www.gnu.org/licenses/>.
+//    along with Seven Update.  If not, see <http://www.gnu.org/licenses/>.
 
 #endregion
 
@@ -80,7 +80,8 @@ namespace SevenUpdate.Pages
         {
             try
             {
-                File.Delete(Shared.UserStore + @"Apps.sul");
+                if (File.Exists(Shared.UserStore + @"Apps.sul"))
+                    File.Delete(Shared.UserStore + @"Apps.sul");
             }
             catch
             {
@@ -89,7 +90,6 @@ namespace SevenUpdate.Pages
             var wc = new WebClient();
 
             wc.DownloadFileCompleted += DownloadFileCompleted;
-
             wc.DownloadFileAsync(SUALocation, Shared.UserStore + @"Apps.sul");
         }
 
@@ -171,7 +171,7 @@ namespace SevenUpdate.Pages
         }
 
         /// <summary>
-        /// Updates the list with the <see cref="userAppList"/>
+        /// Updates the list with the <see cref="userAppList" />
         /// </summary>
         private void UpdateList()
         {
@@ -191,7 +191,7 @@ namespace SevenUpdate.Pages
         }
 
         /// <summary>
-        /// Adds the <see cref="GridViewColumn"/>'s of the <see cref="ListView"/> to be sorted
+        /// Adds the <see cref="GridViewColumn" />'s of the <see cref="ListView" /> to be sorted
         /// </summary>
         private void AddSortBinding()
         {
@@ -306,7 +306,7 @@ namespace SevenUpdate.Pages
         #region ListView Related
 
         /// <summary>
-        /// Updates the <see cref="CollectionView"/> when the <c>userAppList</c> collection changes
+        /// Updates the <see cref="CollectionView" /> when the <c>userAppList</c> collection changes
         /// </summary>
         private void UserAppList_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
         {
@@ -318,7 +318,7 @@ namespace SevenUpdate.Pages
         }
 
         /// <summary>
-        /// Limit the size of the <see cref="GridViewColumn"/> when it's being resized
+        /// Limit the size of the <see cref="GridViewColumn" /> when it's being resized
         /// </summary>
         private void Thumb_DragDelta(object sender, DragDeltaEventArgs e)
         {
@@ -330,7 +330,7 @@ namespace SevenUpdate.Pages
         #region TextBlocks
 
         /// <summary>
-        /// Underlines the text when mouse is over the <see cref="TextBlock"/>
+        /// Underlines the text when mouse is over the <see cref="TextBlock" />
         /// </summary>
         private void TextBlock_MouseEnter(object sender, MouseEventArgs e)
         {
@@ -339,7 +339,7 @@ namespace SevenUpdate.Pages
         }
 
         /// <summary>
-        /// Removes the Underlined text when mouse is leaves the <see cref="TextBlock"/>
+        /// Removes the Underlined text when mouse is leaves the <see cref="TextBlock" />
         /// </summary>
         private void TextBlock_MouseLeave(object sender, MouseEventArgs e)
         {
